pyproject.toml
[project]
name = "madvr-envy-homeassistant"
version = "0.2.1"
description = "Home Assistant integration for madVR Envy"
requires-python = ">=3.11"
dependencies = [
    "madvr-envy>=0.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "ruff>=0.8.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--cov=custom_components/madvr_envy",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
    "-v",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["custom_components/madvr_envy"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]
